% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.compilation.R
\name{count_nodes}
\alias{count_nodes}
\title{Retrives the information from the JAGS compilation}
\usage{
count_nodes(x, ...)
}
\arguments{
\item{x}{A rjags object derived from running \code{jags}.}

\item{...}{Additional parameters that can be passed}
}
\value{
A vector with the number of observed, unobserved and 
total nodes included in the DAG resulting from the model 
assumptions encoded by the code
}
\description{
Retrives the information from the JAGS compilation
}
\examples{
# An example model file is given in:
model.file <- system.file(package="R2jags", "model", "schools.txt")
#=================# 
# initialization  #
#=================#
# data
J <- 8.0
y <- c(28.4,7.9,-2.8,6.8,-0.6,0.6,18.0,12.2)
sd <- c(14.9,10.2,16.3,11.0,9.4,11.4,10.4,17.6)
jags.data <- list("y","sd","J")
jags.params <- c("mu","sigma","theta")
jags.inits <- function(){
 list("mu"=rnorm(1),"sigma"=runif(1),"theta"=rnorm(J))
}
jagsfit <- jags(data=list("y","sd","J"), inits=jags.inits, jags.params,
             n.iter=10, model.file=model.file)
count_nodes(jagsfit)

}
\author{
Gianluca Baio
}
