% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_GB.R
\name{jags_diagplot}
\alias{jags_diagplot}
\title{Specialised diagnostic plots}
\usage{
jags_diagplot(x, what = "Rhat", ...)
}
\arguments{
\item{x}{A 'jags' object resulting from a call to 'jags(...)'}

\item{what}{A string indicating what diagnostic measure should be plotted.
Options are 'Rhat' (default), indicating the PSR statistic, or 'n.eff', 
indicating the effective sample size}

\item{...}{Additional options}
}
\description{
Creates a plot showing the output of convergence indicators, such as
the Potential Scale Reduction and the effective sample size
}
\examples{
\dontrun{ 
} 
}
\seealso{
\code{jags}
}
\author{
Gianluca Baio
}
\keyword{Diagnostic}
\keyword{plots}
