% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_GB.R
\name{jags_summary}
\alias{jags_summary}
\title{Summary table for the output of the MCMC process}
\usage{
jags_summary(x, parameter = NULL, probs = c(0.025, 0.975), ...)
}
\arguments{
\item{x}{A 'jags' object resulting from a call to 'jags(...)'}

\item{parameter}{A (vector of) parameter(s) name(s) to be displayed in the
summary table. If NULL (default), then uses all the model parameters
that have been monitored to construct the summary tibble.}

\item{probs}{A vector of probability values to be shown to represent the 
summary statistics. Defaults to 'c(0.025,0.975)', to show the 2.5 and the 
97.5% percentiles of the posterior distributions (indicating a rough 95%
credible interval). Other possible values are 0.25, 0.5, 0.75}

\item{...}{Additional options}
}
\description{
Creates a summary table based on tibbles and with some nice
formatting properties
}
\examples{
\dontrun{ 
} 
}
\seealso{
\code{jags}, \code{print.jags}
}
\author{
Gianluca Baio
}
\keyword{Summary}
\keyword{table}
