% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autojags.R
\name{autojags}
\alias{autojags}
\alias{update.rjags}
\title{Function for auto-updating \sQuote{JAGS} until the model converges}
\usage{
autojags(
  object,
  n.iter = 1000,
  n.thin = 1,
  Rhat = 1.1,
  n.update = 2,
  refresh = n.iter/50,
  progress.bar,
  ...
)
}
\arguments{
\item{object}{an object of \code{rjags} class.}

\item{n.iter}{number of total iterations per chain, default=1000}

\item{n.thin}{thinning rate. Must be a positive integer, default=1}

\item{Rhat}{converegence criterion, default=1.1.}

\item{n.update}{the max number of updates, default=2.}

\item{refresh}{refresh frequency for progress bar, default is
\code{n.iter/50}}

\item{progress.bar}{type of progress bar. Possible values are
\dQuote{text}, \dQuote{gui}, and \dQuote{none}. Type \dQuote{text} is
displayed on the R console. Type \dQuote{gui} is a graphical progress bar
in a new window. The progress bar is suppressed if \code{progress.bar} is
\dQuote{none}}

\item{\dots}{further arguments pass to or from other methods.}
}
\description{
The \code{autojags} takes a \code{rjags} object as input. \code{autojags}
will update the model until it converges.
}
\examples{


 # see ?jags for an example.


}
\references{
Gelman, A., Carlin, J.B., Stern, H.S., Rubin, D.B. (2003):
\emph{Bayesian Data Analysis}, 2nd edition, CRC Press.
}
\author{
Yu-Sung Su \email{suyusung@tsinghua.edu.cn}
}
\keyword{models}
