\name{recompile}
\alias{recompile}
\alias{recompile.rjags}


\title{Function for recompiling rjags object}
\description{
The \code{recompile} takes a \code{rjags} object as input.  
\code{recompile} will re-compile the previous saved \code{rjags} object.
}

\usage{

recompile(object)
\method{recompile}{rjags}(object)


}


\arguments{
  \item{object}{an object of \code{rjags} class.}
}


\author{ 
  Yu-Sung Su \email{ys463@columbia.edu}
}


\examples{
 # see ?jags for an example.
}

\keyword{models}
